# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rounds
    description: Golf round and side bet operations
  - name: dashboard
    description: Round summaries
  - name: courses
    description: Reusable golf course layouts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /rounds:
    get:
      operationId: listRounds
      tags: [rounds]
      summary: List golf rounds
      responses:
        "200":
          description: Golf rounds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Round"
    post:
      operationId: createRound
      tags: [rounds]
      summary: Create a golf round
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoundInput"
      responses:
        "201":
          description: Created round
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundDetail"
  /rounds/{roundId}:
    get:
      operationId: getRound
      tags: [rounds]
      summary: Get a golf round
      parameters:
        - $ref: "#/components/parameters/RoundId"
      responses:
        "200":
          description: Round detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundDetail"
        "404":
          description: Round not found
    patch:
      operationId: updateRound
      tags: [rounds]
      summary: Update a golf round
      parameters:
        - $ref: "#/components/parameters/RoundId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoundUpdate"
      responses:
        "200":
          description: Updated round
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundDetail"
    delete:
      operationId: deleteRound
      tags: [rounds]
      summary: Delete a golf round
      parameters:
        - $ref: "#/components/parameters/RoundId"
      responses:
        "204":
          description: Round deleted
  /rounds/{roundId}/holes:
    post:
      operationId: recordHole
      tags: [rounds]
      summary: Record a hole result
      parameters:
        - $ref: "#/components/parameters/RoundId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoleResultInput"
      responses:
        "201":
          description: Recorded hole
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HoleResult"
    get:
      operationId: listHoleResults
      tags: [rounds]
      summary: List recorded hole results
      parameters:
        - $ref: "#/components/parameters/RoundId"
      responses:
        "200":
          description: Hole results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HoleResult"
  /rounds/{roundId}/settlement:
    get:
      operationId: getRoundSettlement
      tags: [rounds]
      summary: Get the current round settlement
      parameters:
        - $ref: "#/components/parameters/RoundId"
      responses:
        "200":
          description: Current balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get a summary of tracked rounds
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List saved golf courses
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Saved courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Save a golf course layout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Saved course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a saved golf course
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Saved course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Course not found
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a saved golf course
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Course not found
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a saved golf course
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "204":
          description: Course deleted
  /courses/import-scorecard:
    post:
      operationId: importScorecard
      tags: [courses]
      summary: Extract a course layout from a scorecard image or PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScorecardImportInput"
      responses:
        "200":
          description: Editable extracted course draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseImportDraft"
  /courses/external-search:
    get:
      operationId: searchExternalCourses
      tags: [courses]
      summary: Search an external golf-course provider
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: External search status and results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalCourseSearchResponse"
components:
  parameters:
    RoundId:
      name: roundId
      in: path
      required: true
      schema:
        type: number
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: number
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Player:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        initials:
          type: string
        handicap:
          type: number
        snakeThreshold:
          type: ["number", "null"]
      required: [id, name, initials, handicap, snakeThreshold]
    PlayerInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        handicap:
          type: number
          minimum: 0
          maximum: 54
        snakeThreshold:
          type: ["number", "null"]
          minimum: 2
          maximum: 6
      required: [name]
    DotPoints:
      type: object
      properties:
        greenie:
          type: number
          minimum: 0
        sandy:
          type: number
          minimum: 0
        birdie:
          type: number
          minimum: 0
        eagle:
          type: number
          minimum: 0
        poley:
          type: number
          minimum: 0
        threeputt:
          type: number
          minimum: 0
      required: [greenie, sandy, birdie, eagle, poley, threeputt]
    GameType:
      type: string
      enum: [wolf, nassau, snake, dots]
    Round:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        course:
          type: string
        playedAt:
          type: string
        status:
          type: string
          enum: [in_progress, completed]
        gameTypes:
          type: array
          items:
            $ref: "#/components/schemas/GameType"
        stake:
          type: number
        dollarPerPoint:
          type: number
        wolfUnit:
          type: number
        snakeStake:
          type: number
        dotPoints:
          $ref: "#/components/schemas/DotPoints"
        holePars:
          type: array
          items:
            type: number
        holeStrokeIndex:
          type: array
          items:
            type: number
        currentHole:
          type: number
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
      required:
        [
          id,
          name,
          course,
          playedAt,
          status,
          gameTypes,
          stake,
          dollarPerPoint,
          wolfUnit,
          snakeStake,
          dotPoints,
          holePars,
          holeStrokeIndex,
          currentHole,
          players,
        ]
    RoundInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        course:
          type: string
          minLength: 1
        playedAt:
          type: string
        gameTypes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/GameType"
        stake:
          type: number
          minimum: 0.01
        dollarPerPoint:
          type: number
          minimum: 0
        wolfUnit:
          type: number
          minimum: 0.01
        snakeStake:
          type: number
          minimum: 0
        dotPoints:
          $ref: "#/components/schemas/DotPoints"
        holePars:
          type: array
          minItems: 18
          maxItems: 18
          items:
            type: number
            minimum: 3
            maximum: 6
        holeStrokeIndex:
          type: array
          minItems: 18
          maxItems: 18
          items:
            type: number
            minimum: 1
            maximum: 18
        players:
          type: array
          minItems: 2
          maxItems: 6
          items:
            $ref: "#/components/schemas/PlayerInput"
      required: [name, course, playedAt, gameTypes, stake, players]
    RoundUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        course:
          type: string
          minLength: 1
        status:
          type: string
          enum: [in_progress, completed]
        dollarPerPoint:
          type: number
          minimum: 0
        wolfUnit:
          type: number
          minimum: 0.01
        snakeStake:
          type: number
          minimum: 0
        dotPoints:
          $ref: "#/components/schemas/DotPoints"
        holePars:
          type: array
          minItems: 18
          maxItems: 18
          items:
            type: number
            minimum: 3
            maximum: 6
        holeStrokeIndex:
          type: array
          minItems: 18
          maxItems: 18
          items:
            type: number
            minimum: 1
            maximum: 18
    ScoreInput:
      type: object
      properties:
        playerId:
          type: string
        strokes:
          type: number
          minimum: 1
      required: [playerId, strokes]
    PuttInput:
      type: object
      properties:
        playerId:
          type: string
        putts:
          type: number
          minimum: 0
      required: [playerId, putts]
    DotFlagsInput:
      type: object
      properties:
        playerId:
          type: string
        greenie:
          type: boolean
        sandy:
          type: boolean
        poley:
          type: boolean
      required: [playerId]
    HoleResultInput:
      type: object
      properties:
        hole:
          type: number
          minimum: 1
          maximum: 18
        scores:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/ScoreInput"
        putts:
          type: array
          items:
            $ref: "#/components/schemas/PuttInput"
        wolfPartnerIds:
          type: array
          items:
            type: string
        wolfOverridePlayerId:
          type: ["string", "null"]
        wolfManualResult:
          type: ["string", "null"]
          enum: ["wolfwin", "oppwin", "push", null]
        dots:
          type: array
          items:
            $ref: "#/components/schemas/DotFlagsInput"
        winnerPlayerId:
          type: ["string", "null"]
        snakeHolderPlayerId:
          type: ["string", "null"]
      required: [hole, scores]
    PlayerDots:
      type: object
      properties:
        playerId:
          type: string
        greenie:
          type: boolean
        sandy:
          type: boolean
        poley:
          type: boolean
        birdie:
          type: boolean
        eagle:
          type: boolean
        threeputt:
          type: boolean
      required: [playerId, greenie, sandy, poley, birdie, eagle, threeputt]
    HoleResult:
      allOf:
        - $ref: "#/components/schemas/HoleResultInput"
        - type: object
          properties:
            id:
              type: number
            createdAt:
              type: string
            effectiveWolfPlayerId:
              type: ["string", "null"]
            wolfTeamPlayerIds:
              type: array
              items:
                type: string
            wolfResult:
              type: ["string", "null"]
              enum: ["wolfwin", "oppwin", "push", null]
            wolfCarry:
              type: number
            snakeHolderPlayerId:
              type: ["string", "null"]
            snakeTiePlayerIds:
              type: array
              items:
                type: string
            dotsEarned:
              type: array
              items:
                $ref: "#/components/schemas/PlayerDots"
          required:
            [
              id,
              createdAt,
              effectiveWolfPlayerId,
              wolfTeamPlayerIds,
              wolfResult,
              wolfCarry,
              snakeHolderPlayerId,
              snakeTiePlayerIds,
              dotsEarned,
            ]
    RoundDetail:
      allOf:
        - $ref: "#/components/schemas/Round"
        - type: object
          properties:
            holes:
              type: array
              items:
                $ref: "#/components/schemas/HoleResult"
            settlement:
              $ref: "#/components/schemas/Settlement"
          required: [holes, settlement]
    Balance:
      type: object
      properties:
        playerId:
          type: string
        playerName:
          type: string
        amount:
          type: number
      required: [playerId, playerName, amount]
    Payout:
      type: object
      properties:
        fromPlayerId:
          type: string
        fromPlayerName:
          type: string
        toPlayerId:
          type: string
        toPlayerName:
          type: string
        amount:
          type: number
      required: [fromPlayerId, fromPlayerName, toPlayerId, toPlayerName, amount]
    PlayerPoints:
      type: object
      properties:
        playerId:
          type: string
        playerName:
          type: string
        wolfPoints:
          type: number
        dotsPoints:
          type: number
        snakePoints:
          type: number
        nassauFrontAmount:
          type: number
        nassauBackAmount:
          type: number
        nassauOverallAmount:
          type: number
        nassauAmount:
          type: number
      required:
        [
          playerId,
          playerName,
          wolfPoints,
          dotsPoints,
          snakePoints,
          nassauFrontAmount,
          nassauBackAmount,
          nassauOverallAmount,
          nassauAmount,
        ]
    Settlement:
      type: object
      properties:
        balances:
          type: array
          items:
            $ref: "#/components/schemas/Balance"
        payouts:
          type: array
          items:
            $ref: "#/components/schemas/Payout"
        pointTotals:
          type: array
          items:
            $ref: "#/components/schemas/PlayerPoints"
        snakeHolderPlayerId:
          type: ["string", "null"]
        holesRecorded:
          type: number
        totalPot:
          type: number
      required: [balances, payouts, pointTotals, snakeHolderPlayerId, holesRecorded, totalPot]
    DashboardSummary:
      type: object
      properties:
        totalRounds:
          type: number
        activeRounds:
          type: number
        totalHolesRecorded:
          type: number
        latestRound:
          anyOf:
            - $ref: "#/components/schemas/Round"
            - type: "null"
      required: [totalRounds, activeRounds, totalHolesRecorded, latestRound]
    CourseHole:
      type: object
      properties:
        hole:
          type: number
          minimum: 1
          maximum: 18
        par:
          type: number
          minimum: 3
          maximum: 6
        strokeIndex:
          type: number
          minimum: 1
          maximum: 18
      required: [hole, par, strokeIndex]
    Course:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        location:
          type: string
        holes:
          type: array
          minItems: 18
          maxItems: 18
          items:
            $ref: "#/components/schemas/CourseHole"
        source:
          type: string
          enum: [manual, upload, external]
        sourceDocumentName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, location, holes, source, sourceDocumentName, createdAt, updatedAt]
    CourseInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
        holes:
          type: array
          minItems: 18
          maxItems: 18
          items:
            $ref: "#/components/schemas/CourseHole"
        source:
          type: string
          enum: [manual, upload, external]
        sourceDocumentName:
          type: ["string", "null"]
      required: [name, holes]
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
        holes:
          type: array
          minItems: 18
          maxItems: 18
          items:
            $ref: "#/components/schemas/CourseHole"
    ScorecardImportInput:
      type: object
      properties:
        fileName:
          type: string
          minLength: 1
        mimeType:
          type: string
          enum: [application/pdf, image/jpeg, image/png, image/webp]
        data:
          type: string
          minLength: 1
      required: [fileName, mimeType, data]
    CourseImportDraft:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        holes:
          type: array
          minItems: 18
          maxItems: 18
          items:
            $ref: "#/components/schemas/CourseHole"
        source:
          type: string
          enum: [upload]
        sourceDocumentName:
          type: string
        warnings:
          type: array
          items:
            type: string
        confidence:
          type: ["number", "null"]
      required: [name, location, holes, source, sourceDocumentName, warnings, confidence]
    ExternalCourse:
      type: object
      properties:
        externalId:
          type: string
        name:
          type: string
        location:
          type: string
        holes:
          type: array
          items:
            $ref: "#/components/schemas/CourseHole"
      required: [externalId, name, location, holes]
    ExternalCourseSearchResponse:
      type: object
      properties:
        available:
          type: boolean
        provider:
          type: ["string", "null"]
        message:
          type: string
        courses:
          type: array
          items:
            $ref: "#/components/schemas/ExternalCourse"
      required: [available, provider, message, courses]