# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourseHole } from './courseHole';
import type { CourseSource } from './courseSource';

export interface Course {
  id: number;
  name: string;
  location: string;
  /**
     * @minItems 18
     * @maxItems 18
     */
  holes: CourseHole[];
  source: CourseSource;
  /** @nullable */
  sourceDocumentName: string | null;
  createdAt: string;
  updatedAt: string;
}