# lib/api-zod/src/generated/types/courseHole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseHole {
  /**
     * @minimum 1
     * @maximum 18
     */
  hole: number;
  /**
     * @minimum 3
     * @maximum 6
     */
  par: number;
  /**
     * @minimum 1
     * @maximum 18
     */
  strokeIndex: number;
}