# lib/api-zod/src/generated/types/courseImportDraft.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourseHole } from './courseHole';
import type { CourseImportDraftSource } from './courseImportDraftSource';

export interface CourseImportDraft {
  name: string;
  location: string;
  /**
     * @minItems 18
     * @maxItems 18
     */
  holes: CourseHole[];
  source: CourseImportDraftSource;
  sourceDocumentName: string;
  warnings: string[];
  /** @nullable */
  confidence: number | null;
}