# lib/api-zod/src/generated/types/courseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourseHole } from './courseHole';
import type { CourseInputSource } from './courseInputSource';

export interface CourseInput {
  /** @minLength 1 */
  name: string;
  location?: string;
  /**
     * @minItems 18
     * @maxItems 18
     */
  holes: CourseHole[];
  source?: CourseInputSource;
  /** @nullable */
  sourceDocumentName?: string | null;
}