# lib/api-zod/src/generated/types/courseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourseHole } from './courseHole';

export interface CourseUpdate {
  /** @minLength 1 */
  name?: string;
  location?: string;
  /**
     * @minItems 18
     * @maxItems 18
     */
  holes?: CourseHole[];
}