# lib/api-zod/src/generated/types/externalCourse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourseHole } from './courseHole';

export interface ExternalCourse {
  externalId: string;
  name: string;
  location: string;
  holes: CourseHole[];
}