# lib/api-zod/src/generated/types/externalCourseSearchResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExternalCourse } from './externalCourse';

export interface ExternalCourseSearchResponse {
  available: boolean;
  /** @nullable */
  provider: string | null;
  message: string;
  courses: ExternalCourse[];
}