# lib/api-zod/src/generated/types/scorecardImportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScorecardImportInputMimeType } from './scorecardImportInputMimeType';

export interface ScorecardImportInput {
  /** @minLength 1 */
  fileName: string;
  mimeType: ScorecardImportInputMimeType;
  /** @minLength 1 */
  data: string;
}